+++ src/hooks/z_hook.c
@@ -6,7 +6,7 @@
 /*   By: vicgarci <vicgarci@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/22 18:49:25 by vicgarci          #+#    #+#             */
-/*   Updated: 2023/02/23 16:32:44 by vicgarci         ###   ########.fr       */
+/*   Updated: 2023/02/27 14:49:46 by vicgarci         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	z_hook(t_FdF_info *fdf)
 {
-	if (fdf->cam->dim.z != (INT_MIN + 1))
+	if ((int)fdf->cam->dim.z != (INT_MIN + 1))
 	{
 		fdf->cam->dim.z -= 0.1f;
 		ft_printf("Valor de la cota: %d\n", (int)fdf->cam->dim.z);
